Update to support Julia v0.5
This update adds support for Julia v0.5.  The main change is that SvgRay
is now inside of a module.  This is still a bit cumbersome, as local
directory is not in the julia path.  However, it encapsulates the code
and makes it easier to work with.

The other changes are mainly to do with vector arithmetic.  A few
vectors are transposed or had their dimensions and rank altered.